% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odmx_main.R
\name{getLabValuesGreaterOrEqualZero}
\alias{getLabValuesGreaterOrEqualZero}
\title{Get lab results from database}
\usage{
getLabValuesGreaterOrEqualZero(
  db,
  stringsAsFactors = FALSE,
  keep.only.best.quality = TRUE
)
}
\arguments{
\item{db}{name of ODBC database or full path to database file}

\item{stringsAsFactors}{Passed to \code{hsSqlQuery}. Default: \code{FALSE}#'}

\item{keep.only.best.quality}{if \code{TRUE} (default) data values in raw
data quality are removed if they occur in different quality levels}
}
\description{
Get lab results from database
}
