% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odmx_main.R
\name{keepOnlyBestQualityControlLevel}
\alias{keepOnlyBestQualityControlLevel}
\title{Remove raw data values if data values of better quality exist}
\usage{
keepOnlyBestQualityControlLevel(x, dbg = TRUE)
}
\arguments{
\item{x}{data frame with columns \code{ValueID}, \code{LabSampleCode},
\code{VariableCode}, \code{QualityControlLevelID}}

\item{dbg}{if \code{TRUE} (default) debug messages are shown else not}
}
\description{
Remove raw data values if data values of better quality exist
}
