% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odmx_main.R
\name{simplifyColumns}
\alias{simplifyColumns}
\title{merge VariableCode and Unit, use '<', '>' or '=' as CensorCode}
\usage{
simplifyColumns(x, short.codes = TRUE)
}
\arguments{
\item{x}{data frame with columns \code{VariableCode, UnitsAbbreviation,
CensorCode}}

\item{short.codes}{if \code{TRUE} (default) the censor codes are renamed:
"lt" -> "<"; "gt" -> ">"; "nc" -> ""}
}
\description{
merge VariableCode and Unit, use '<', '>' or '=' as CensorCode
}
\examples{
x <- data.frame(VariableCode = c("A", "B"),
                UnitsAbbreviation = c("mg/l", "ug/l"),
                CensorCode = c("lt", "nc"))

simplifyColumns(x)
}
